feat: NotificationLogActivity에 bottom sheet 기능 병합
