--- android/app/src/main/java/com/woowacourse/ody/presentation/meetinglog/CodeCopyListener.kt
@@ -1,5 +1,0 @@
-package com.woowacourse.ody.presentation.meetinglog
-
-interface CodeCopyListener {
-    fun copy()
-}
