--- android/app/src/main/java/com/woowacourse/ody/presentation/meetinglog/LogsListener.kt
@@ -1,5 +1,0 @@
-package com.woowacourse.ody.presentation.meetinglog
-
-interface LogsListener {
-    fun share()
-}
