--- android/app/src/main/java/com/woowacourse/ody/presentation/notificationlog/CopyInviteCodeButtonListener.kt
@@ -1,5 +1,0 @@
-package com.woowacourse.ody.presentation.notificationlog
-
-interface CopyInviteCodeButtonListener {
-    fun onClickCopyInviteCode()
-}
