+++ android/app/src/main/java/com/woowacourse/ody/presentation/notificationlog/NotificationLogActivity.kt
@@ -1,19 +1,24 @@
 package com.woowacourse.ody.presentation.notificationlog
 
+import android.content.ClipData
+import android.content.ClipboardManager
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.annotation.StringRes
 import androidx.appcompat.app.AppCompatActivity
+import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.snackbar.Snackbar
+import com.woowacourse.ody.R
 import com.woowacourse.ody.data.local.fake.FakeMeetingRepository
 import com.woowacourse.ody.data.local.fake.FakeNotificationLogRepository
 import com.woowacourse.ody.databinding.ActivityNotificationLogBinding
 import com.woowacourse.ody.presentation.notificationlog.adapter.NotificationLogListAdapter
 
-class NotificationLogActivity : AppCompatActivity() {
+class NotificationLogActivity : AppCompatActivity(), CopyInviteCodeListener, ShareListener {
     private val viewModel: NotificationLogViewModel by viewModels {
         NotificationLogViewModelFactory(
             // DefaultNotificationLogRepository,
@@ -28,14 +33,18 @@ class NotificationLogActivity : AppCompatActivity() {
 
     private val adapter: NotificationLogListAdapter by lazy {
         NotificationLogListAdapter(
-            viewModel,
+            this,
         )
     }
+
+    private val bottomSheetLayout by lazy { findViewById<ConstraintLayout>(R.id.cl_bottom_sheet) }
+    private lateinit var bottomSheetBehavior: BottomSheetBehavior<ConstraintLayout>
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         initializeBinding()
         initializeObserve()
+        initializePersistentBottomSheet()
         viewModel.initialize()
     }
 
@@ -44,6 +53,8 @@ class NotificationLogActivity : AppCompatActivity() {
         binding.lifecycleOwner = this
         binding.rvNotificationLog.adapter = adapter
         binding.rvNotificationLog.layoutManager = LinearLayoutManager(this)
+        binding.shareListener = this
+        binding.copyInviteCodeListener = this
         setContentView(binding.root)
     }
 
@@ -53,6 +64,24 @@ class NotificationLogActivity : AppCompatActivity() {
         }
     }
 
+    private fun initializePersistentBottomSheet() {
+        bottomSheetBehavior = BottomSheetBehavior.from(bottomSheetLayout)
+    }
+
+    override fun share() {
+        bottomSheetBehavior.state = BottomSheetBehavior.STATE_EXPANDED
+    }
+
+    override fun copyInviteCode() {
+        bottomSheetBehavior.state = BottomSheetBehavior.STATE_HIDDEN
+        val inviteCode = viewModel.meeting.value?.inviteCode
+        val clipboard = getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
+        val clip = ClipData.newPlainText(INVITE_CODE_LABEL, inviteCode)
+        clipboard.setPrimaryClip(clip)
+
+        showSnackBar(R.string.logs_copy_code_guide)
+    }
+
     private fun showSnackBar(
         @StringRes messageId: Int,
     ) = Snackbar.make(binding.root, messageId, Snackbar.LENGTH_SHORT)
@@ -60,6 +89,8 @@ class NotificationLogActivity : AppCompatActivity() {
         .show()
 
     companion object {
+        private const val INVITE_CODE_LABEL = "inviteCode"
+
         fun getIntent(context: Context): Intent = Intent(context, NotificationLogActivity::class.java)
     }
 }
