+++ android/app/src/main/java/com/woowacourse/ody/presentation/notificationlog/ShareListener.kt
@@ -1,0 +1,5 @@
+package com.woowacourse.ody.presentation.notificationlog
+
+interface ShareListener {
+    fun share()
+}
