--- android/app/src/main/res/layout/activity_logs.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="logsListener"
-            type="com.woowacourse.ody.presentation.meetinglog.LogsListener" />
-
-        <variable
-            name="codeCopyListener"
-            type="com.woowacourse.ody.presentation.meetinglog.CodeCopyListener" />
-    </data>
-
-    <androidx.coordinatorlayout.widget.CoordinatorLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".presentation.meetinglog.LogsActivity">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <ImageView
-                android:id="@+id/iv_share"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="18dp"
-                android:layout_marginEnd="18dp"
-                android:onClick="@{() -> logsListener.share()}"
-                android:src="@drawable/ic_share"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-        <include
-            layout="@layout/bottom_sheet_code_copy"
-            app:codeCopyListener="@{codeCopyListener}" />
-
-    </androidx.coordinatorlayout.widget.CoordinatorLayout>
-</layout>
-
