+++ android/app/src/main/res/layout/toolbar_log.xml
@@ -7,6 +7,9 @@
         <variable
             name="title"
             type="java.lang.String" />
+        <variable
+            name="shareListener"
+            type="com.woowacourse.ody.presentation.notificationlog.ShareListener" />
     </data>
 
     <androidx.appcompat.widget.Toolbar
@@ -48,6 +51,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginHorizontal="14dp"
                 android:src="@drawable/ic_share"
+                android:onClick="@{() -> shareListener.share()}"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toEndOf="@id/tv_title"
